requirements.txt
anthropic==0.86.0
ccxt==4.5.45
hmmlearn==0.3.3
joblib==1.5.2
numpy==2.3.2
pandas==2.3.2
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
websockets>=12.0
